+++ projects/frontend/es-app/src/views/gamelist/ISimpleGameListView.cpp
@@ -67,19 +67,6 @@ void ISimpleGameListView::onFileChanged(FileData* file, FileChangeType change) {
 		return ;
 	}
 
-	int index = getCursorIndex();
-
-	if (change == FileChangeType::FILE_SORTED) {
-        index = 0; // go back to the top level
-    	// When sorting, cursor is reset to the root one, so, we need to flush the cursor stack
-		while (!mCursorStack.empty()) {
-			mCursorStack.pop();
-		}
-	}
-
-	populateList(getRoot()->getChildren());
-	setCursorIndex(index);
-
     if (change == FileChangeType::FILE_REMOVED) {
         bool favorite = file->metadata.get("favorite") == "true";
         delete file;
@@ -88,6 +75,17 @@ void ISimpleGameListView::onFileChanged(FileData* file, FileChangeType change) {
             ViewController::get()->getSystemListView()->manageFavorite();
         }
     }
+
+	FileData* cursor = getCursor();
+
+	if (RecalboxConf::getInstance()->getBool(getRoot()->getSystem()->getName() + ".flatfolder")) {
+		populateList(getRoot());
+	} else {
+		refreshList();
+	}
+
+	setCursor(cursor);
+
 
 	/* Favorite */
 	if (file->getType() == GAME) {
@@ -130,7 +128,7 @@ bool ISimpleGameListView::input(InputConfig* config, Input input) {
 				// it's a folder
 				if (cursor->getChildren().size() > 0) {
 					mCursorStack.push(cursor);
-					populateList(cursor->getChildren());
+					populateList(cursor);
 					setCursorIndex(0);
 				}
 			}
@@ -145,7 +143,7 @@ bool ISimpleGameListView::input(InputConfig* config, Input input) {
 				mCursorStack.pop();
 
 				FileData* cursor = mCursorStack.size() ? mCursorStack.top() : getRoot();
-				populateList(cursor->getChildren());
+				populateList(cursor);
 
 				setCursor(selected);
 				//Sound::getFromTheme(getTheme(), getName(), "back")->play();
@@ -192,11 +190,7 @@ bool ISimpleGameListView::input(InputConfig* config, Input input) {
 
 					// Reload to refresh the favorite icon
 					int cursorPlace = getCursorIndex();
-					if (RecalboxConf::getInstance()->getBool(getCursor()->getSystem()->getName() + ".flatfolder")) {
-						populateList(getRoot()->getChildren());
-					} else {
-						populateList(cursor->getParent()->getChildren());
-					}
+                    refreshList();
 					setCursorIndex(cursorPlace + (removeFavorite ? -1 : 1));
 				}
 			}